Move import candidates filters to own file
